hotfix: rename hooks, eslint fix
