+++ src/navigation/tabs/discover/discover-tab.model.ts
@@ -2,7 +2,6 @@ export type DiscoverTabParamsList = {
   DiscoverScreen: undefined;
 };
 
-export enum DISCOVER_STACK_ROTES  {
+export enum DISCOVER_STACK_ROTES {
   DiscoverScreen = 'DiscoverScreen'
-};
-
+}
