+++ src/navigation/tabs/home/home-tab.model.ts
@@ -2,7 +2,6 @@ export type HomeTabParamsList = {
   HomeScreen: undefined;
 };
 
-export enum HOME_STACK_ROTES  {
+export enum HOME_STACK_ROTES {
   HomeScreen = 'HomeScreen'
-};
-
+}
