+++ src/navigation/tabs/index.ts
@@ -1,8 +1,5 @@
 export * from './discover/discover-stack';
-export * from './discover/routes';
 
 export * from './home/home-stack';
-export * from './home/routes';
 
 export * from './settings/setting-stack';
-export * from './settings/routes';
