+++ src/navigation/tabs/settings/settings-tab.model.ts
@@ -2,7 +2,6 @@ export type SettingsTabParamsList = {
   SettingsScreen: undefined;
 };
 
-export enum SETTINGS_STACK_ROTES  {
+export enum SETTINGS_STACK_ROTES {
   SettingsScreen = 'SettingsScreen'
-};
-
+}
