+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './svgs';
 export * from './localization';
-export * from './navigation';
